Move all the form on a folder
